Update bootbox.js
Added the ability for options.message to be a jQuery object, when creating a dialog. The jQuery object must have the class "bootbox-template". This will clone the object, including methods, and remove the 'hidden' class. The reasoning is that you can make a pretty complicated forms, with events attached to them and whatnot. This allows you to create them in your standard html and just give them the classes "hidden" and "bootbox-template", and pass that into bootbox as the message.