Add defaultAction for prompt
defaultAction is set to confirm by default. This defaultAction is used when submitting the form (i.e. pressing Enter). Before, bootbox used the 'btn-primary' class to find the button, then click it. However, this caused a problem if this was used on a different button than the default button. In my case, I had 'btn-primary' for 'Cancel' and 'btn-success' for 'Confirm' meaning the 'Cancel' button would get pressed when hitting Enter, which is confusing to the user.